requirements.txt

scapy==2.5.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.7.1
PyYAML==6.0.1
joblib==1.3.2
Flask==2.3.2
gunicorn==25.0.1
colorama==0.4.6
pytest==7.4.0